Authorization system created but does not confirm token.

+++ Core/TalepYonetimi.Application/Handlers/CommandHandlers/Demands/DeleteDemandCommandHandler.cs
@@ -1,5 +1,7 @@
 ﻿using AutoMapper;
 using MediatR;
+using Microsoft.AspNetCore.Identity;
+using Microsoft.EntityFrameworkCore;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -7,6 +9,7 @@ using System.Text;
 using System.Threading.Tasks;
 using TalepYonetimi.Application.AbstractRepositories.Demands;
 using TalepYonetimi.Application.Commands.Demands;
+using TalepYonetimi.Domain.Entities.Admin;
 
 namespace TalepYonetimi.Application.Handlers.CommandHandlers.Demands
 {
@@ -14,16 +17,27 @@ namespace TalepYonetimi.Application.Handlers.CommandHandlers.Demands
     {
         private readonly IDemandReadRepository demandReadRepository;
         private readonly IDemandWriteRepository demandWriteRepository;
-        public DeleteDemandCommandHandler(IDemandReadRepository _demandReadRepository, IDemandWriteRepository _demandWriteRepository)
+        private readonly UserManager<ApplicationUser> userManager; 
+        public DeleteDemandCommandHandler(IDemandReadRepository _demandReadRepository, IDemandWriteRepository _demandWriteRepository, UserManager<ApplicationUser> _userManager)
         {
             demandReadRepository = _demandReadRepository;
             demandWriteRepository = _demandWriteRepository;
+            userManager = _userManager;
         }
         public async Task<bool> Handle(DeleteDemandCommand request, CancellationToken cancellationToken)
         {
-            var demand = await demandReadRepository.GetByIdAsync(request.Id);
+            var user = await userManager.Users.FirstOrDefaultAsync(u => u.Id == request.UserId);
 
-            if (demand != null)
+            if (user == null)
+            {
+                return false;
+            }
+
+            var departmentId = user.Department.Id;
+
+            var demand = await demandReadRepository.GetByIdAsync(request.DemandId);
+
+            if (demand != null && demand.Department.Id == departmentId)
             {
                 await demandWriteRepository.RemoveAsync(demand);
                 return await demandWriteRepository.SaveChangesAsync();
